Adding header into root page also (for player profile)
